Fix missed dep changes in TestUtils move
